configurando a barra de pesquisa
